pyproject.toml
[tool.poetry]
name = "inception-bot"
version = "0.1.0"
description = "inception-bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugin_dirs = ["inception_bot/plugins"]
plugins = ["nonebot_plugin_apscheduler"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
